分布式演示

+++ src/main/java/com/ujigu/secure/demo/algorithm/PreciseModuloTableShardingAlgorithm.java
@@ -17,15 +17,15 @@
 
 package com.ujigu.secure.demo.algorithm;
 
+import java.util.Collection;
+
 import io.shardingjdbc.core.api.algorithm.sharding.PreciseShardingValue;
 import io.shardingjdbc.core.api.algorithm.sharding.standard.PreciseShardingAlgorithm;
 
-import java.util.Collection;
-
-public final class PreciseModuloTableShardingAlgorithm implements PreciseShardingAlgorithm<Long> {
+public final class PreciseModuloTableShardingAlgorithm implements PreciseShardingAlgorithm<Integer> {
     
     @Override
-    public String doSharding(final Collection<String> availableTargetNames, final PreciseShardingValue<Long> shardingValue) {
+    public String doSharding(final Collection<String> availableTargetNames, final PreciseShardingValue<Integer> shardingValue) {
         for (String each : availableTargetNames) {
             if (each.endsWith(shardingValue.getValue() % 2 + "")) {
                 return each;
